feat: formatDateString 모듈 추가 #22
- Dashboard와 RequestResult에서 사용되는 date format 함수 모듈화
